+++ scripts/git-subtree/typewriter/pull.ps1
@@ -1,0 +1,4 @@
+git subtree pull `
+    -P workspace/plugins/typewriter `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-typewriter.git `
+    dev